+++ src/pages/tu/teaching/courses/components/Step3Review.tsx
@@ -1,0 +1,265 @@
+/**
+ * Step 3: 검토 및 저장
+ * 담당: 최종 검토 화면
+ */
+import {
+  Upload,
+  Plus,
+  Pencil,
+  CheckCircle2,
+  AlertTriangle,
+  BookOpen,
+  Calendar,
+  Tag,
+  Globe,
+  FileText,
+} from 'lucide-react';
+import { Button, Label, Card, CardHeader, CardContent, Alert, AlertDescription } from '@/components/common';
+import type { CourseFormData } from '@/types';
+import type { CategoryResponse } from '@/types/common';
+import { translations, levelOptions, type TranslationKey } from './courseCreate.constants';
+
+interface Step3ReviewProps {
+  language: 'ko' | 'en';
+  formData: CourseFormData;
+  categories: CategoryResponse[];
+  onGoToStep: (step: number) => void;
+}
+
+export function Step3Review({
+  language,
+  formData,
+  categories,
+  onGoToStep,
+}: Readonly<Step3ReviewProps>) {
+  const getText = (key: TranslationKey) =>
+    language === 'ko' ? translations[key].ko : translations[key].en;
+
+  // 경고 메시지 계산
+  const warnings: string[] = [];
+  if (!formData.title) warnings.push(getText('warningCourse'));
+  if (!formData.categoryId) warnings.push(getText('warningCategory'));
+  if (formData.lessons.length === 0) warnings.push(getText('warningLesson'));
+
+  return (
+    <div className="flex flex-col gap-6">
+      {/* 헤더 */}
+      <div>
+        <h2 className="text-text-primary mb-2">{getText('reviewTitle')}</h2>
+        <p className="text-text-secondary m-0">{getText('reviewDesc')}</p>
+      </div>
+
+      {/* 경고 메시지 또는 완료 메시지 */}
+      {warnings.length > 0 ? (
+        <Alert variant="destructive">
+          <AlertTriangle size={16} />
+          <AlertDescription>
+            <strong>{getText('warningTitle')}</strong>
+            <ul className="mt-2 mb-0 pl-4">
+              {warnings.map((warning, index) => (
+                <li key={index}>{warning}</li>
+              ))}
+            </ul>
+          </AlertDescription>
+        </Alert>
+      ) : (
+        <Alert variant="info">
+          <CheckCircle2 size={16} />
+          <AlertDescription>
+            <strong>{getText('readyToSubmit')}</strong>
+            <p className="mt-1 mb-0">{getText('readyToSubmitDesc')}</p>
+          </AlertDescription>
+        </Alert>
+      )}
+
+      {/* 기본 정보 섹션 */}
+      <Card>
+        <CardHeader className="pb-4">
+          <div className="flex items-center justify-between">
+            <div className="flex items-center gap-3">
+              <BookOpen size={20} className="text-action-primary" />
+              <h3 className="text-text-primary m-0 text-base font-medium">{getText('basicInfo')}</h3>
+            </div>
+            <Button variant="ghost" size="sm" onClick={() => onGoToStep(1)} className="border border-border">
+              <Pencil size={14} />
+              {getText('edit')}
+            </Button>
+          </div>
+        </CardHeader>
+        <CardContent className="pt-0">
+          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+            {/* 강의명 */}
+            <div className="md:col-span-2">
+              <Label className="text-text-secondary text-sm">{getText('courseName')}</Label>
+              <p className="text-text-primary mt-1 mb-0">
+                {formData.title || <span className="text-text-tertiary italic">{getText('notEntered')}</span>}
+              </p>
+            </div>
+
+            {/* 강의 설명 */}
+            <div className="md:col-span-2">
+              <Label className="text-text-secondary text-sm">{getText('courseDescription')}</Label>
+              <p className="text-text-primary mt-1 mb-0 whitespace-pre-wrap">
+                {formData.description || (
+                  <span className="text-text-tertiary italic">{getText('notEntered')}</span>
+                )}
+              </p>
+            </div>
+
+            {/* 카테고리 */}
+            <div>
+              <Label className="text-text-secondary text-sm">{getText('category')}</Label>
+              <p className="text-text-primary mt-1 mb-0">
+                {formData.categoryId ? (
+                  categories.find((cat) => cat.id === formData.categoryId)?.name
+                ) : (
+                  <span className="text-text-tertiary italic">{getText('notEntered')}</span>
+                )}
+              </p>
+            </div>
+
+            {/* 난이도 */}
+            <div>
+              <Label className="text-text-secondary text-sm">{getText('difficulty')}</Label>
+              <p className="text-text-primary mt-1 mb-0">
+                {formData.level ? (
+                  levelOptions.find((opt) => opt.value === formData.level)?.label
+                ) : (
+                  <span className="text-text-tertiary italic">{getText('notEntered')}</span>
+                )}
+              </p>
+            </div>
+
+            {/* 수강 기간 */}
+            <div className="md:col-span-2">
+              <Label className="text-text-secondary text-sm flex items-center gap-1">
+                <Calendar size={14} />
+                {getText('period')}
+              </Label>
+              <p className="text-text-primary mt-1 mb-0">
+                {formData.startDate && formData.endDate ? (
+                  `${formData.startDate} ~ ${formData.endDate}`
+                ) : formData.startDate ? (
+                  `${formData.startDate} ~`
+                ) : formData.endDate ? (
+                  `~ ${formData.endDate}`
+                ) : (
+                  <span className="text-text-tertiary italic">{getText('noPeriod')}</span>
+                )}
+              </p>
+            </div>
+
+            {/* 태그 */}
+            <div className="md:col-span-2">
+              <Label className="text-text-secondary text-sm flex items-center gap-1">
+                <Tag size={14} />
+                {getText('tags')}
+              </Label>
+              <div className="mt-2">
+                {formData.tags.length > 0 ? (
+                  <div className="flex flex-wrap gap-2">
+                    {formData.tags.map((tag, index) => (
+                      <span
+                        key={index}
+                        className="px-2.5 py-1 bg-bg-secondary text-text-primary text-sm rounded-md"
+                      >
+                        {tag}
+                      </span>
+                    ))}
+                  </div>
+                ) : (
+                  <span className="text-text-tertiary italic">{getText('noTags')}</span>
+                )}
+              </div>
+            </div>
+
+            {/* 다국어 설정 */}
+            <div className="md:col-span-2">
+              <Label className="text-text-secondary text-sm flex items-center gap-1">
+                <Globe size={14} />
+                {getText('multiLanguage')}
+              </Label>
+              <p className="text-text-primary mt-1 mb-0">
+                {formData.multiLanguage.enabled ? (
+                  <>
+                    <span className="text-status-success">{getText('enabled')}</span>
+                    {formData.multiLanguage.languages.length > 0 && (
+                      <span className="text-text-secondary ml-2">
+                        ({formData.multiLanguage.languages.length}
+                        {getText('languageCount')}: {formData.multiLanguage.languages.map((l) => l.name).join(', ')})
+                      </span>
+                    )}
+                  </>
+                ) : (
+                  <span className="text-text-tertiary">{getText('disabled')}</span>
+                )}
+              </p>
+            </div>
+          </div>
+        </CardContent>
+      </Card>
+
+      {/* 커리큘럼 섹션 */}
+      <Card>
+        <CardHeader className="pb-4">
+          <div className="flex items-center justify-between">
+            <div className="flex items-center gap-3">
+              <FileText size={20} className="text-action-primary" />
+              <h3 className="text-text-primary m-0 text-base font-medium">{getText('curriculum')}</h3>
+              {formData.lessons.length > 0 && (
+                <span className="text-text-secondary text-sm">
+                  ({getText('totalLessons')}: {formData.lessons.length}, {getText('totalContents')}:{' '}
+                  {formData.lessons.reduce((acc, lesson) => acc + lesson.contents.length, 0)})
+                </span>
+              )}
+            </div>
+            <Button variant="ghost" size="sm" onClick={() => onGoToStep(2)} className="border border-border">
+              <Pencil size={14} />
+              {getText('edit')}
+            </Button>
+          </div>
+        </CardHeader>
+        <CardContent className="pt-0">
+          {formData.lessons.length > 0 ? (
+            <div className="flex flex-col gap-3">
+              {formData.lessons.map((lesson) => (
+                <div key={lesson.id} className="p-4 bg-bg-secondary rounded-lg flex items-start gap-3">
+                  <div className="w-7 h-7 rounded-md bg-btn-neutral text-white flex items-center justify-center font-medium text-sm shrink-0">
+                    {lesson.order}
+                  </div>
+                  <div className="flex-1 min-w-0">
+                    <p className="text-text-primary font-medium m-0">
+                      {lesson.title || `${getText('lessonNumber')} ${lesson.order}`}
+                    </p>
+                    {lesson.description && (
+                      <p className="text-text-secondary text-sm mt-1 mb-0 line-clamp-2">
+                        {lesson.description}
+                      </p>
+                    )}
+                    {lesson.contents.length > 0 && (
+                      <div className="flex items-center gap-2 mt-2">
+                        <Upload size={14} className="text-text-tertiary" />
+                        <span className="text-text-secondary text-sm">
+                          {lesson.contents.length}
+                          {getText('contentsCount')}
+                        </span>
+                      </div>
+                    )}
+                  </div>
+                </div>
+              ))}
+            </div>
+          ) : (
+            <div className="text-center py-8">
+              <p className="text-text-secondary m-0">{getText('noCurriculum')}</p>
+              <Button variant="ghost" size="sm" onClick={() => onGoToStep(2)} className="mt-3 border border-border">
+                <Plus size={14} />
+                {getText('addLesson')}
+              </Button>
+            </div>
+          )}
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
