+++ src/pages/tu/teaching/courses/components/courseCreate.constants.ts
@@ -1,0 +1,81 @@
+/**
+ * CourseCreate 페이지 공통 상수
+ */
+
+export const translations = {
+  title: { ko: '강의 등록', en: 'Create Course' },
+  loadTemplate: { ko: '템플릿 불러오기', en: 'Load Template' },
+  close: { ko: '닫기', en: 'Close' },
+  previous: { ko: '이전', en: 'Previous' },
+  next: { ko: '다음', en: 'Next' },
+  saveDraft: { ko: '임시저장', en: 'Save Draft' },
+  submit: { ko: '강의 등록', en: 'Submit' },
+  lastSaved: { ko: '마지막 저장', en: 'Last saved' },
+  step1: { ko: '기본 정보', en: 'Basic Info' },
+  step2: { ko: '회차 구성', en: 'Curriculum' },
+  step3: { ko: '검토 및 저장', en: 'Review' },
+  courseName: { ko: '강의명', en: 'Course Name' },
+  courseNamePlaceholder: { ko: '강의명을 입력하세요', en: 'Enter course name' },
+  courseDescription: { ko: '강의 설명', en: 'Course Description' },
+  courseDescriptionPlaceholder: { ko: '강의에 대한 설명을 입력하세요', en: 'Enter course description' },
+  category: { ko: '카테고리', en: 'Category' },
+  selectCategory: { ko: '카테고리 선택', en: 'Select category' },
+  difficulty: { ko: '난이도', en: 'Difficulty' },
+  selectDifficulty: { ko: '난이도 선택', en: 'Select difficulty' },
+  beginner: { ko: '입문', en: 'Beginner' },
+  elementary: { ko: '초급', en: 'Elementary' },
+  intermediate: { ko: '중급', en: 'Intermediate' },
+  advanced: { ko: '고급', en: 'Advanced' },
+  startDate: { ko: '시작일', en: 'Start Date' },
+  endDate: { ko: '종료일', en: 'End Date' },
+  comingSoon: { ko: '준비 중입니다', en: 'Coming Soon' },
+  curriculumTitle: { ko: '회차 구성', en: 'Curriculum' },
+  curriculumDesc: { ko: '회차를 추가하고 콘텐츠를 등록하세요. 드래그앤드롭으로 순서를 변경할 수 있습니다.', en: 'Add lessons and register content. Drag and drop to reorder.' },
+  noLessons: { ko: '아직 등록된 회차가 없습니다.', en: 'No lessons registered yet.' },
+  addFirstLesson: { ko: '첫 번째 회차 추가', en: 'Add First Lesson' },
+  addLesson: { ko: '회차 추가', en: 'Add Lesson' },
+  lessonTitle: { ko: '회차 제목', en: 'Lesson Title' },
+  lessonTitlePlaceholder: { ko: '회차 제목을 입력하세요', en: 'Enter lesson title' },
+  lessonDescription: { ko: '회차 설명', en: 'Lesson Description' },
+  lessonDescriptionPlaceholder: { ko: '회차에 대한 설명을 입력하세요', en: 'Enter lesson description' },
+  contents: { ko: '콘텐츠', en: 'Contents' },
+  noContents: { ko: '등록된 콘텐츠가 없습니다.', en: 'No content registered.' },
+  fileUpload: { ko: '파일 업로드', en: 'File Upload' },
+  externalLink: { ko: '외부 링크', en: 'External Link' },
+  loadExisting: { ko: '기존 콘텐츠 불러오기', en: 'Load Existing Content' },
+  deleteLesson: { ko: '회차 삭제', en: 'Delete Lesson' },
+  reviewTitle: { ko: '강의 정보 검토', en: 'Review Course Information' },
+  reviewDesc: { ko: '등록할 강의 정보를 확인하세요. 수정이 필요하면 각 섹션의 수정 버튼을 눌러주세요.', en: 'Please review your course information. Click edit to modify each section.' },
+  edit: { ko: '수정', en: 'Edit' },
+  basicInfo: { ko: '기본 정보', en: 'Basic Information' },
+  tags: { ko: '태그', en: 'Tags' },
+  noTags: { ko: '태그 없음', en: 'No tags' },
+  period: { ko: '수강 기간', en: 'Course Period' },
+  noPeriod: { ko: '기간 미설정', en: 'Not set' },
+  multiLanguage: { ko: '다국어 설정', en: 'Multi-language' },
+  enabled: { ko: '활성화', en: 'Enabled' },
+  disabled: { ko: '비활성화', en: 'Disabled' },
+  languageCount: { ko: '개 언어', en: ' languages' },
+  curriculum: { ko: '커리큘럼', en: 'Curriculum' },
+  totalLessons: { ko: '총 회차', en: 'Total lessons' },
+  totalContents: { ko: '총 콘텐츠', en: 'Total contents' },
+  lessonNumber: { ko: '회차', en: 'Lesson' },
+  contentsCount: { ko: '개 콘텐츠', en: ' contents' },
+  noCurriculum: { ko: '등록된 회차가 없습니다.', en: 'No lessons registered.' },
+  notEntered: { ko: '미입력', en: 'Not entered' },
+  warningTitle: { ko: '입력 확인 필요', en: 'Input Required' },
+  warningCourse: { ko: '강의명이 입력되지 않았습니다.', en: 'Course name is required.' },
+  warningCategory: { ko: '카테고리가 선택되지 않았습니다.', en: 'Category is required.' },
+  warningLesson: { ko: '최소 1개의 회차가 필요합니다.', en: 'At least one lesson is required.' },
+  readyToSubmit: { ko: '강의 등록 준비 완료', en: 'Ready to Submit' },
+  readyToSubmitDesc: { ko: '모든 필수 정보가 입력되었습니다. 강의 등록 버튼을 눌러 강의를 등록하세요.', en: 'All required information has been entered. Click Submit to register your course.' },
+} as const;
+
+export type TranslationKey = keyof typeof translations;
+
+export const levelOptions = [
+  { value: '', label: '난이도 선택' },
+  { value: 'BEGINNER', label: '초급' },
+  { value: 'INTERMEDIATE', label: '중급' },
+  { value: 'ADVANCED', label: '고급' },
+];
