+++ src/pages/tu/teaching/courses/components/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Course Create Page 컴포넌트 모음
+ * 담당자별 역할 분리:
+ * - Step1BasicInfo: 강의 기본 정보 (course 테이블)
+ * - Step2Curriculum + LessonCard: 회차/콘텐츠 구성 (content 테이블)
+ * - Step3Review: 최종 검토
+ */
+export { Step1BasicInfo } from './Step1BasicInfo';
+export { Step2Curriculum } from './Step2Curriculum';
+export { Step3Review } from './Step3Review';
+export { LessonCard } from './LessonCard';
+export { translations, levelOptions } from './courseCreate.constants';
+export type { TranslationKey } from './courseCreate.constants';
